requirements.txt
flask
flask-cors
pandas
yfinance
fredapi
python-dotenv
certifi
gunicorn